+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 * Fix #417: Support Subresources
 * Fix #1548: Allow user to update the status on CustomResources
 * Fix #1282: Add Support for fetching Kubernetes metrics
+* Fix #1917 Allow user to add custom headers to client
 
 ### 4.6.4 (20-11-2019)
 #### Bugs